requirements.txt
streamlit
streamlit_extras
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
pillow
googletrans